Перенес код связанный с работой с событиями в отдельный presenter.
